% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubLocViz.R
\name{subLocViz}
\alias{subLocViz}
\title{Sub-cellular visualization}
\usage{
subLocViz(
  gene_name,
  stage = c("D0", "D3", "D5"),
  use_data = c("scaled", "ratio", "raw"),
  plot_type = c("point", "density", "hexagonal", "buble"),
  color_by = NULL,
  show_gene_name = FALSE,
  label_size = 2,
  point_size = 1,
  point_shape = 16,
  continuous_colors = continuous_colorbar(),
  text_size_circle = 3,
  text_color_circle = "purple",
  ...
)
}
\arguments{
\item{gene_name}{Gene symbols used for visualization}

\item{stage}{Stages used for visualization}

\item{use_data}{use normalized data?}

\item{plot_type}{type of plot?}

\item{color_by}{which column used for visualization?}

\item{show_gene_name}{Whether to show gene names?}

\item{label_size}{The gene names size in figure}

\item{point_size}{The point size}

\item{point_shape}{The point shape}

\item{continuous_colors}{The color of points}

\item{text_size_circle}{the text color of outlier}

\item{text_color_circle}{the color of outlier}

\item{...}{other parameters.}
}
\value{
a ggplot2 object
}
\description{
Sub-cellular visualization
}
